Remove dead code and test-only wrappers
