+++ internal/mouse/test_helpers_test.go
@@ -1,0 +1,5 @@
+package mouse
+
+func (p *Parser) InProgress() bool {
+	return p.state != stateNone
+}
