+++ test/hooks_test.go
@@ -8,100 +8,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/fsnotify/fsnotify"
 	"github.com/weill-labs/amux/internal/server"
 )
-
-func waitForFile(t *testing.T, path string, timeout time.Duration) bool {
-	t.Helper()
-
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		t.Fatalf("new watcher: %v", err)
-	}
-	defer watcher.Close()
-
-	dir := filepath.Dir(path)
-	if err := os.MkdirAll(dir, 0700); err != nil {
-		t.Fatalf("mkdir %s: %v", dir, err)
-	}
-	if err := watcher.Add(dir); err != nil {
-		t.Fatalf("watch %s: %v", dir, err)
-	}
-	if _, err := os.Stat(path); err == nil {
-		return true
-	}
-
-	timer := time.NewTimer(timeout)
-	defer timer.Stop()
-	ticker := time.NewTicker(50 * time.Millisecond)
-	defer ticker.Stop()
-	for {
-		if _, err := os.Stat(path); err == nil {
-			return true
-		}
-		select {
-		case <-timer.C:
-			return false
-		case <-ticker.C:
-		case event := <-watcher.Events:
-			if filepath.Clean(event.Name) != filepath.Clean(path) {
-				continue
-			}
-			if _, err := os.Stat(path); err == nil {
-				return true
-			}
-		case err := <-watcher.Errors:
-			t.Logf("fsnotify error for %s: %v", path, err)
-		}
-	}
-}
-
-func waitForFileContent(t *testing.T, path string, timeout time.Duration) string {
-	t.Helper()
-
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		t.Fatalf("new watcher: %v", err)
-	}
-	defer watcher.Close()
-
-	dir := filepath.Dir(path)
-	if err := os.MkdirAll(dir, 0700); err != nil {
-		t.Fatalf("mkdir %s: %v", dir, err)
-	}
-	if err := watcher.Add(dir); err != nil {
-		t.Fatalf("watch %s: %v", dir, err)
-	}
-	if data, err := os.ReadFile(path); err == nil && len(data) > 0 {
-		return string(data)
-	}
-
-	timer := time.NewTimer(timeout)
-	defer timer.Stop()
-	ticker := time.NewTicker(50 * time.Millisecond)
-	defer ticker.Stop()
-	for {
-		if data, err := os.ReadFile(path); err == nil && len(data) > 0 {
-			return string(data)
-		}
-		select {
-		case <-timer.C:
-			return ""
-		case <-ticker.C:
-		case event := <-watcher.Events:
-			if filepath.Clean(event.Name) != filepath.Clean(path) {
-				continue
-			}
-			data, err := os.ReadFile(path)
-			if err == nil && len(data) > 0 {
-				return string(data)
-			}
-		case err := <-watcher.Errors:
-			t.Logf("fsnotify error for %s: %v", path, err)
-		}
-	}
-}
 
 func waitForHook(t *testing.T, h *ServerHarness, event, pane string, after uint64) {
 	t.Helper()
